[#1] refactor: usecase 통합
