+++ src/main/java/com/baedal/rider/adapter/presentation/controller/RiderController.java
@@ -3,8 +3,7 @@ package com.baedal.rider.adapter.presentation.controller;
 import com.baedal.rider.adapter.presentation.request.LoginRequest;
 import com.baedal.rider.adapter.presentation.request.SignupRequest;
 import com.baedal.rider.adapter.presentation.response.LoginResponse;
-import com.baedal.rider.application.port.in.RiderAuthenticationUsecase;
-import com.baedal.rider.application.port.in.RiderSignupUsecase;
+import com.baedal.rider.application.port.in.RiderUsecase;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -17,19 +16,18 @@ import org.springframework.web.bind.annotation.RestController;
 @Slf4j
 public class RiderController {
 
-  private final RiderAuthenticationUsecase authenticationUsecase;
-  private final RiderSignupUsecase signupUsecase;
+  private final RiderUsecase riderUsecase;
 
   @PostMapping("/v0/login")
   public ResponseEntity<LoginResponse> login(@RequestBody LoginRequest request) {
-    LoginResponse response = authenticationUsecase.authenticate(request.email(),
+    LoginResponse response = riderUsecase.authenticate(request.email(),
         request.password());
     return ResponseEntity.ok(response);
   }
 
   @PostMapping("/v0/signup")
   public ResponseEntity<Void> singup(@RequestBody SignupRequest request) {
-    signupUsecase.signup(request.email(), request.nickname(), request.password());
+    riderUsecase.signup(request.email(), request.nickname(), request.password());
     return ResponseEntity.noContent().build();
   }
 }
