--- src/main/java/com/baedal/rider/application/port/in/RiderSignupUsecase.java
@@ -1,6 +1,0 @@
-package com.baedal.rider.application.port.in;
-
-public interface RiderSignupUsecase {
-
-  public void signup(String email, String nickname, String rawPassword);
-}
