+++ src/main/java/com/baedal/rider/application/service/RiderService.java
@@ -1,6 +1,7 @@
 package com.baedal.rider.application.service;
 
-import com.baedal.rider.application.port.in.RiderSignupUsecase;
+import com.baedal.rider.adapter.presentation.response.LoginResponse;
+import com.baedal.rider.application.port.in.RiderUsecase;
 import com.baedal.rider.domain.entity.Rider;
 import com.baedal.rider.domain.repository.RiderRepository;
 import lombok.RequiredArgsConstructor;
@@ -10,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class RiderSignupService implements RiderSignupUsecase {
+public class RiderService implements RiderUsecase {
 
   private final RiderRepository repository;
   private final PasswordEncoder passwordEncoder;
@@ -19,5 +20,21 @@ public class RiderSignupService implements RiderSignupUsecase {
   public void signup(String email, String nickname, String rawPassword) {
     Rider rider = new Rider(email, nickname, passwordEncoder.encode(rawPassword));
     repository.save(rider);
+  }
+
+  @Override
+  public LoginResponse authenticate(String email, String password) {
+    // TODO: Authentication Exception 정의
+    Rider rider = repository.findByEmail(email)
+        .orElseThrow(() -> new RuntimeException("Wrong email or password"));
+
+    if (!passwordEncoder.matches(password, rider.getPassword())) {
+      throw new RuntimeException("Wrong email or password");
+    }
+
+    return new LoginResponse(
+        rider.getId(),
+        rider.getEmail()
+    );
   }
 }
