+++ src/main/resources/application-local.yml
@@ -1,0 +1,2 @@
+server:
+  port: 8081
