+++ src/main/resources/application.yml
@@ -5,3 +5,4 @@ spring:
   output:
     ansi:
       enabled: always
+
